Move logic into hook
